seperated roll call and election handling
