+++ be1-go/hub/organizer.go
@@ -636,97 +636,6 @@ func (c *laoChannel) processRollCallObject(msg message.Message) error {
 	return nil
 }
 
-func (c *laoChannel) processCreateRollCall(data message.Data) error {
-	rollCallData := data.(*message.CreateRollCallData)
-	if !c.checkRollCallID(rollCallData.Creation, message.Stringer(rollCallData.Name), rollCallData.ID) {
-		return &message.Error{
-			Code:        -4,
-			Description: "The id of the roll call does not correspond to SHA256(‘R’||lao_id||creation||name)",
-		}
-	}
-
-	c.rollCall.id = string(rollCallData.ID)
-	c.rollCall.state = Created
-	return nil
-}
-
-func (c *laoChannel) processOpenRollCall(data message.Data, action message.RollCallAction) error {
-	if action == message.RollCallAction(message.OpenRollCallAction) {
-		// If the action is an OpenRollCallAction,
-		// the previous roll call action should be a CreateRollCallAction
-		if c.rollCall.state != Created {
-			return &message.Error{
-				Code:        -1,
-				Description: "The roll call can not be opened since it has not been created previously",
-			}
-		}
-	} else {
-		// If the action is an RepenRollCallAction,
-		// the previous roll call action should be a CloseRollCallAction
-		if c.rollCall.state != Closed {
-			return &message.Error{
-				Code:        -1,
-				Description: "The roll call can not be reopened since it has not been closed previously",
-			}
-		}
-	}
-
-	rollCallData := data.(*message.OpenRollCallData)
-
-	if !c.rollCall.checkPrevID(rollCallData.Opens) {
-		return &message.Error{
-			Code:        -4,
-			Description: "The field `opens` does not correspond to the id of the previous roll call message",
-		}
-	}
-
-	opens := base64.StdEncoding.EncodeToString(rollCallData.Opens)
-	if !c.checkRollCallID(message.Stringer(opens), rollCallData.OpenedAt, rollCallData.UpdateID) {
-		return &message.Error{
-			Code:        -4,
-			Description: "The id of the roll call does not correspond to SHA256(‘R’||lao_id||opens||opened_at)",
-		}
-	}
-
-	c.rollCall.id = string(rollCallData.UpdateID)
-	c.rollCall.state = Open
-	return nil
-}
-
-func (c *laoChannel) processCloseRollCall(data message.Data) error {
-	if c.rollCall.state != Open {
-		return &message.Error{
-			Code:        -1,
-			Description: "The roll call can not be closed since it is not open",
-		}
-	}
-
-	rollCallData := data.(*message.CloseRollCallData)
-	if !c.rollCall.checkPrevID(rollCallData.Closes) {
-		return &message.Error{
-			Code:        -4,
-			Description: "The field `closes` does not correspond to the id of the previous roll call message",
-		}
-	}
-
-	closes := base64.StdEncoding.EncodeToString(rollCallData.Closes)
-	if !c.checkRollCallID(message.Stringer(closes), rollCallData.ClosedAt, rollCallData.UpdateID) {
-		return &message.Error{
-			Code:        -4,
-			Description: "The id of the roll call does not correspond to SHA256(‘R’||lao_id||closes||closed_at)",
-		}
-	}
-
-	c.rollCall.id = string(rollCallData.UpdateID)
-	c.rollCall.state = Closed
-	c.attendees = map[string]struct{}{}
-	for i := 0; i < len(rollCallData.Attendees); i += 1 {
-		c.attendees[string(rollCallData.Attendees[i])] = struct{}{}
-	}
-
-	return nil
-}
-
 func (c *laoChannel) processElectionObject(msg message.Message) error {
 	action := message.ElectionAction(msg.Data.GetAction())
 
@@ -743,84 +652,4 @@ func (c *laoChannel) processElectionObject(msg message.Message) error {
 	}
 
 	return nil
-}
-
-func (c *laoChannel) createElection(msg message.Message) error {
-	organizerHub := c.hub
-
-	organizerHub.Lock()
-	defer organizerHub.Unlock()
-
-	// Check the data
-	data, ok := msg.Data.(*message.ElectionSetupData)
-	if !ok {
-		return &message.Error{
-			Code:        -4,
-			Description: "failed to cast data to SetupElectionData",
-		}
-	}
-
-	// Check if the Lao ID of the message corresponds to the channel ID
-	encodedLaoID := base64.URLEncoding.EncodeToString(data.LaoID)
-	channelID := c.channelID[6:]
-	if channelID != encodedLaoID {
-		return &message.Error{
-			Code:        -4,
-			Description: fmt.Sprintf("Lao ID of the message (Lao: %s) is different from the channelID (channel: %s)", encodedLaoID, channelID),
-		}
-	}
-
-	// Compute the new election channel id
-	encodedElectionID := base64.URLEncoding.EncodeToString(data.ID)
-	encodedID := encodedLaoID + "/" + encodedElectionID
-
-	// Create the new election channel
-	electionCh := electionChannel{
-		createBaseChannel(organizerHub, rootPrefix+encodedID),
-		data.StartTime,
-		data.EndTime,
-		false,
-		getAllQuestionsForElectionChannel(data.Questions, data),
-	}
-
-	// Add the SetupElection message to the new election channel
-	messageID := base64.URLEncoding.EncodeToString(msg.MessageID)
-	electionCh.inboxMu.Lock()
-	electionCh.inbox[messageID] = msg
-	electionCh.inboxMu.Unlock()
-
-	// Add the new election channel to the organizerHub
-	organizerHub.channelByID[encodedID] = &electionCh
-
-	return nil
-}
-
-func getAllQuestionsForElectionChannel(questions []message.Question, data *message.ElectionSetupData) map[string]question {
-	qs := make(map[string]question)
-	for _, q := range questions {
-		qs[base64.URLEncoding.EncodeToString(q.ID)] = question{
-			q.ID,
-			q.BallotOptions,
-			sync.RWMutex{},
-			make(map[string]validVote),
-			q.VotingMethod,
-		}
-	}
-	return qs
-}
-
-func (r *rollCall) checkPrevID(prevID []byte) bool {
-	return string(prevID) == r.id
-}
-
-// Check if the id of the roll call corresponds to the hash of the correct parameters
-// Return true if the hash corresponds to the id and false otherwise
-func (c *laoChannel) checkRollCallID(str1, str2 fmt.Stringer, id []byte) bool {
-	laoID := c.channelID[6:]
-	hash, err := message.Hash(message.Stringer('R'), message.Stringer(laoID), str1, str2)
-	if err != nil {
-		return false
-	}
-
-	return bytes.Equal(hash, id)
 }
